项目整理完成：准备GitHub上传
✨ 主要改进：
- 🗂️ 完整重组文件结构：按功能分类组织测试文件
- 🚮 删除所有debug文件：移除临时调试代码，保持项目整洁
- 📝 更新文档：创建完整的README.md和文件组织说明
- 🔧 修复编译问题：解决类型警告和编译错误
- 🧹 清理重复文件：删除空文件和功能重复的测试

📁 最终文件结构：
- hashmap.hpp - 核心实现（中文注释）
- test/basic/ - 基础功能测试
- test/functional/ - 功能性测试
- test/stl_compliance/ - STL兼容性测试
- test/hash/ - 哈希函数测试
- utils/ - 工具库文件
- README.md - 完整项目文档

✅ 项目状态：
- 所有测试通过
- 内存安全验证完成
- CMake构建配置优化
- .gitignore配置完善
- 准备好GitHub上传

+++ test/functional/test_expansion.cpp
@@ -16,12 +16,12 @@ int main() {
         std::cout << "  Will expand after inserting " << threshold << " elements\n\n";
         
         // Insert elements up to and beyond the threshold
-        for (int i = 0; i < threshold + 5; ++i) {
+        for (size_t i = 0; i < threshold + 5; ++i) {
             std::cout << "Inserting " << i << "...\n";
             
             size_t old_bucket_count = map.bucket_count();
             
-            auto result = map.insert(i, i * 10);
+            auto result = map.insert(static_cast<int>(i), static_cast<int>(i * 10));
             
             std::cout << "  Size: " << map.size() 
                       << ", Load factor: " << map.load_factor()
